Changed the output stack from 32 to 16-bit

+++ video_aligner.py
@@ -342,10 +342,11 @@ if len(transform_mat != 0):  # If I have a matrix either from file or calculated
             ],
             axis=1,
         )  # Save stacked g and irm image
+        stacked_video = stacked_video.astype(np.uint16)
 
     tifffile.imwrite(
         output_path + Path(wt_path).stem + "_multichannel_aligned.tif",
-        np.float32(stacked_video),
+        stacked_video,
         imagej=True,
         metadata={
             "Composite mode": "composite",  # This is what was needed for fiji to open it merged
